Update
When the user presses Enter and a process is activated, the start datetime will be updated with the moment when that action occurs.

When they use the "Stop" button, the start datetime will be cleared to "--/--/-- --:--."

When the user keeps the cursor inside the "Current Setpoint," the background will turn white; when they exit, it will return to gray.

When the user uses the arrow keys to increase or decrease the setpoint, if the "Current Setpoint" is "--" or empty, it will default to 5.

Limit the number of characters in the "Current Setpoint" to 2.

Limit the possibility of not increasing beyond 85 with the arrow keys and not decreasing below 5.